// backend/src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
      applications: {
        Row: {
          address: string
          admission_number: string
          application_id: string
          course_id: string | null
          course_name: string
          created_at: string
          department: string
          email: string
          full_name: string
          gender: string
          id: string
          parent_name: string
          parent_phone: string
          phone: string
          photo_url: string | null
          roll_number: string
          semester: string
          user_id: string | null
        }
        Insert: {
          address?: string
          admission_number: string
          application_id: string
          course_id?: string | null
          course_name: string
          created_at?: string
          department: string
          email: string
          full_name: string
          gender: string
          id?: string
          parent_name?: string
          parent_phone?: string
          phone: string
          photo_url?: string | null
          roll_number: string
          semester: string
          user_id?: string | null
        }
        Update: {
          address?: string
          admission_number?: string
          application_id?: string
          course_id?: string | null
          course_name?: string
          created_at?: string
          department?: string
          email?: string
          full_name?: string
          gender?: string
          id?: string
          parent_name?: string
          parent_phone?: string
          phone?: string
          photo_url?: string | null
          roll_number?: string
          semester?: string
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "applications_course_id_fkey"
            columns: ["course_id"]
            isOneToOne: false
            referencedRelation: "courses"
            referencedColumns: ["id"]
          },
        ]
      }
      courses: {
        Row: {
          created_at: string
          description: string
          duration: string
          fee: number
          id: string
          image_url: string | null
          name: string
          seats: number
          start_date: string | null
          status: string
          trainer: string
          updated_at: string
        }
        Insert: {
          created_at?: string
          description?: string
          duration?: string
          fee?: number
          id?: string
          image_url?: string | null
          name: string
          seats?: number
          start_date?: string | null
          status?: string
          trainer?: string
          updated_at?: string
        }
        Update: {
          created_at?: string
          description?: string
          duration?: string
          fee?: number
          id?: string
          image_url?: string | null
          name?: string
          seats?: number
          start_date?: string | null
          status?: string
          trainer?: string
          updated_at?: string
        }
        Relationships: []
      }
      user_roles: {
        Row: {
          created_at: string
          id: string
          role: Database["public"]["Enums"]["app_role"]
          user_id: string
        }
        Insert: {
          created_at?: string
          id?: string
          role: Database["public"]["Enums"]["app_role"]
          user_id: string
        }
        Update: {
          created_at?: string
          id?: string
          role?: Database["public"]["Enums"]["app_role"]
          user_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      has_role: {
        Args: {
          _role: Database["public"]["Enums"]["app_role"]
          _user_id: string
        }
        Returns: boolean
      }
    }
    Enums: {
      app_role: "admin"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never